requirements.txt
langchain
langchain-community
langchain-openai
openai
streamlit
python-dotenv
faiss-cpu
pypdf
python-docx
beautifulsoup4
tiktoken
langchain-google-genai
google-generativeai 